requirements.txt
Flask
Flask_SQLAlchemy
psycopg2-binary
Flask-WTF==1.1.1
Werkzeug==2.2.3
python-dotenv